Refactored project folder structure
