+++ .gitignore
@@ -100,3 +100,5 @@ ENV/
 .ropeproject
 
 *.log
+
+.vscode/*
