+++ app/models.py
@@ -44,7 +44,7 @@ from sqlalchemy import Index, UniqueConstraint
 from sqlalchemy.orm import relationship, backref, sessionmaker
 from flask_sqlalchemy import Model, SQLAlchemy
 from psycopg2 import OperationalError
-from connection import get_database_uri
+from app.connection import get_database_uri
 
 """ Base DB Model """
 class BaseModel(Model):
