+++ app/static/js/jquery.js
