+++ app/templates/base.html
@@ -61,9 +61,17 @@
       <ul class="sidebar-menu" data-widget="tree">
         <li class="header">Administration</li>
         <!-- Optionally, you can add icons to the links -->
-        <li class="active"><a href="/recommendations/manage"><i class="fa fa-link"></i> <span>Manage Recommendation</span></a></li>
+        <li class="active"><a href="/recommendations/manage">
+          <i class="fa fa-link"></i> <span>Manage Recommendation</span></a>
+          <ul>
+            <li>Create</li>
+            <li>List</li>
+            <li>Get</li>
+            <li>Update</li>
+          </ul>
+        </li>
         <li><a href="/recommendations/metadata"><i class="fa fa-link"></i> <span>Manage Product Metadata</span></a></li>
-        <li><a href="/recommendations/docs"><i class="fa fa-link"></i> <span>Documenation</span></a></li>
+        <li><a href="/recommendations/docs"><i class="fa fa-link"></i> <span>Documentation</span></a></li>
       </ul>
       <!-- /.sidebar-menu -->
     </section>
