+++ setup.cfg
@@ -5,4 +5,4 @@ spec-color=1
 nologcapture=1
 with-coverage=1
 cover-erase=1
-cover-package=server,models,engine
+cover-package=app
